Extract Msf::DBManager::Import::Appscan
MSP-11124
